Struct aws_sdk_fsx::client::fluent_builders::CreateFileSystemFromBackup

pub struct CreateFileSystemFromBackup<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> CreateFileSystemFromBackup<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<CreateFileSystemFromBackupOutput, SdkError<CreateFileSystemFromBackupError>> where
R::Policy: SmithyRetryPolicy<CreateFileSystemFromBackupInputOperationOutputAlias, CreateFileSystemFromBackupOutput, CreateFileSystemFromBackupError, CreateFileSystemFromBackupInputOperationRetryAlias>,

pub fn backup_id(self, inp: impl Into<String>) -> Self

The ID of the source backup. Specifies the backup you are copying.

pub fn set_backup_id(self, input: Option<String>) -> Self

pub fn client_request_token(self, inp: impl Into<String>) -> Self

A string of up to 64 ASCII characters that Amazon FSx uses to ensure idempotent creation. This string is automatically filled on your behalf when you use the Command Line Interface (CLI) or an Amazon Web Services SDK.

pub fn set_client_request_token(self, input: Option<String>) -> Self

pub fn subnet_ids(self, inp: impl Into<String>) -> Self

Appends an item to SubnetIds.

To override the contents of this collection use set_subnet_ids.

Specifies the IDs of the subnets that the file system will be accessible from. For Windows MULTI_AZ_1 file system deployment types, provide exactly two subnet IDs, one for the preferred file server and one for the standby file server. You specify one of these subnets as the preferred subnet using the WindowsConfiguration > PreferredSubnetID property.

For Windows SINGLE_AZ_1 and SINGLE_AZ_2 deployment types and Lustre file systems, provide exactly one subnet ID. The file server is launched in that subnet's Availability Zone.

pub fn set_subnet_ids(self, input: Option<Vec<String>>) -> Self

pub fn security_group_ids(self, inp: impl Into<String>) -> Self

Appends an item to SecurityGroupIds.

To override the contents of this collection use set_security_group_ids.

A list of IDs for the security groups that apply to the specified network interfaces created for file system access. These security groups apply to all network interfaces. This value isn't returned in later DescribeFileSystem requests.

pub fn set_security_group_ids(self, input: Option<Vec<String>>) -> Self

pub fn tags(self, inp: impl Into<Tag>) -> Self

Appends an item to Tags.

To override the contents of this collection use set_tags.

The tags to be applied to the file system at file system creation. The key value of the Name tag appears in the console as the file system name.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn windows_configuration(
self,
inp: CreateFileSystemWindowsConfiguration
) -> Self

The configuration for this Microsoft Windows file system.

pub fn set_windows_configuration(
self,
input: Option<CreateFileSystemWindowsConfiguration>
) -> Self

pub fn lustre_configuration(
self,
inp: CreateFileSystemLustreConfiguration
) -> Self

The Lustre configuration for the file system being created.

pub fn set_lustre_configuration(
self,
input: Option<CreateFileSystemLustreConfiguration>
) -> Self

pub fn storage_type(self, inp: StorageType) -> Self

Sets the storage type for the Windows file system you're creating from a backup. Valid values are SSD and HDD.

Set to SSD to use solid state drive storage. Supported on all Windows deployment types.

Set to HDD to use hard disk drive storage. Supported on SINGLE_AZ_2 and MULTI_AZ_1 Windows file system deployment types.

Default value is SSD.

HDD and SSD storage types have different minimum storage capacity requirements. A restored file system's storage capacity is tied to the file system that was backed up. You can create a file system that uses HDD storage from a backup of a file system that used SSD storage only if the original SSD file system had a storage capacity of at least 2000 GiB.

pub fn set_storage_type(self, input: Option<StorageType>) -> Self

pub fn kms_key_id(self, inp: impl Into<String>) -> Self

The ID of the Key Management Service (KMS) key used to encrypt the file system's data for Amazon FSx for Windows File Server file systems, Amazon FSx for NetApp ONTAP file systems, and Amazon FSx for Lustre PERSISTENT_1 file systems at rest. If not specified, the Amazon FSx managed key is used. The Amazon FSx for Lustre SCRATCH_1 and SCRATCH_2 file systems are always encrypted at rest using Amazon FSx managed keys. For more information, see Encrypt in the Key Management Service API Reference.

pub fn set_kms_key_id(self, input: Option<String>) -> Self