Module aws_sdk_fsx::error

Modules

active_directory_error

See ActiveDirectoryError

backup_being_copied

See BackupBeingCopied

backup_in_progress

See BackupInProgress

backup_not_found

See BackupNotFound

backup_restoring

See BackupRestoring

bad_request

See BadRequest

data_repository_task_ended

See DataRepositoryTaskEnded

data_repository_task_executing

See DataRepositoryTaskExecuting

data_repository_task_not_found

See DataRepositoryTaskNotFound

file_system_not_found

See FileSystemNotFound

incompatible_parameter_error

See IncompatibleParameterError

incompatible_region_for_multi_az

See IncompatibleRegionForMultiAz

internal_server_error

See InternalServerError

invalid_destination_kms_key

See InvalidDestinationKmsKey

invalid_export_path

See InvalidExportPath

invalid_import_path

See InvalidImportPath

invalid_network_settings

See InvalidNetworkSettings

invalid_per_unit_storage_throughput

See InvalidPerUnitStorageThroughput

invalid_region

See InvalidRegion

invalid_source_kms_key

See InvalidSourceKmsKey

missing_file_system_configuration

See MissingFileSystemConfiguration

missing_volume_configuration

See MissingVolumeConfiguration

not_service_resource_error

See NotServiceResourceError

resource_does_not_support_tagging

See ResourceDoesNotSupportTagging

resource_not_found

See ResourceNotFound

service_limit_exceeded

See ServiceLimitExceeded

source_backup_unavailable

See SourceBackupUnavailable

storage_virtual_machine_not_found

See StorageVirtualMachineNotFound

unsupported_operation

See UnsupportedOperation

volume_not_found

See VolumeNotFound

Structs

ActiveDirectoryError

An Active Directory error.

AssociateFileSystemAliasesError

BackupBeingCopied

You can't delete a backup while it's being copied.

BackupInProgress

Another backup is already under way. Wait for completion before initiating

BackupNotFound

No Amazon FSx backups were found based upon the supplied parameters.

BackupRestoring

You can't delete a backup while it's being used to restore a file

BadRequest

A generic error indicating a failure with a client request.

CancelDataRepositoryTaskError

CopyBackupError

CreateBackupError

CreateDataRepositoryTaskError

CreateFileSystemError

CreateFileSystemFromBackupError

CreateStorageVirtualMachineError

CreateVolumeError

CreateVolumeFromBackupError

DataRepositoryTaskEnded

The data repository task could not be canceled because the task has already ended.

DataRepositoryTaskExecuting

An existing data repository task is currently executing on the file system.

DataRepositoryTaskNotFound

The data repository task or tasks you specified could not be found.

DeleteBackupError

DeleteFileSystemError

DeleteStorageVirtualMachineError

DeleteVolumeError

DescribeBackupsError

DescribeDataRepositoryTasksError

DescribeFileSystemAliasesError

DescribeFileSystemsError

DescribeStorageVirtualMachinesError

DescribeVolumesError

DisassociateFileSystemAliasesError

FileSystemNotFound

No Amazon FSx file systems were found based upon supplied parameters.

IncompatibleParameterError

The error returned when a second request is received with the same client request

IncompatibleRegionForMultiAz

Amazon FSx doesn't support Multi-AZ Windows File Server

InternalServerError

A generic error indicating a server-side failure.

InvalidDestinationKmsKey

The Key Management Service (KMS) key of the destination

InvalidExportPath

The path provided for data repository export isn't valid.

InvalidImportPath

The path provided for data repository import isn't valid.

InvalidNetworkSettings

One or more network settings specified in the request are invalid.

InvalidPerUnitStorageThroughput

An invalid value for PerUnitStorageThroughput was provided. Please create your file system again, using a valid value.

InvalidRegion

The Region provided for Source Region is invalid or

InvalidSourceKmsKey

The Key Management Service (KMS) key of the source backup

ListTagsForResourceError

MissingFileSystemConfiguration

A file system configuration is required for this operation.

MissingVolumeConfiguration

A volume configuration is required for this operation.

NotServiceResourceError

The resource specified for the tagging operation is not a resource type owned by

ResourceDoesNotSupportTagging

The resource specified does not support tagging.

ResourceNotFound

The resource specified by the Amazon Resource Name (ARN) can't be found.

ServiceLimitExceeded

An error indicating that a particular service limit was exceeded. You can increase

SourceBackupUnavailable

The request was rejected because the lifecycle status of the

StorageVirtualMachineNotFound

No Amazon FSx for NetApp ONTAP SVMs were found based upon the supplied parameters.

TagResourceError

UnsupportedOperation

The requested operation is not supported for this resource or API.

UntagResourceError

UpdateFileSystemError

UpdateStorageVirtualMachineError

UpdateVolumeError

VolumeNotFound

No Amazon FSx for NetApp ONTAP volumes were found based upon the supplied parameters.

Enums

AssociateFileSystemAliasesErrorKind

CancelDataRepositoryTaskErrorKind

CopyBackupErrorKind

CreateBackupErrorKind

CreateDataRepositoryTaskErrorKind

CreateFileSystemErrorKind

CreateFileSystemFromBackupErrorKind

CreateStorageVirtualMachineErrorKind

CreateVolumeErrorKind

CreateVolumeFromBackupErrorKind

DeleteBackupErrorKind

DeleteFileSystemErrorKind

DeleteStorageVirtualMachineErrorKind

DeleteVolumeErrorKind

DescribeBackupsErrorKind

DescribeDataRepositoryTasksErrorKind

DescribeFileSystemAliasesErrorKind

DescribeFileSystemsErrorKind

DescribeStorageVirtualMachinesErrorKind

DescribeVolumesErrorKind

DisassociateFileSystemAliasesErrorKind

ListTagsForResourceErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateFileSystemErrorKind

UpdateStorageVirtualMachineErrorKind

UpdateVolumeErrorKind