Module aws_sdk_fsx::input

Modules

associate_file_system_aliases_input

See AssociateFileSystemAliasesInput

cancel_data_repository_task_input

See CancelDataRepositoryTaskInput

copy_backup_input

See CopyBackupInput

create_backup_input

See CreateBackupInput

create_data_repository_task_input

See CreateDataRepositoryTaskInput

create_file_system_from_backup_input

See CreateFileSystemFromBackupInput

create_file_system_input

See CreateFileSystemInput

create_storage_virtual_machine_input

See CreateStorageVirtualMachineInput

create_volume_from_backup_input

See CreateVolumeFromBackupInput

create_volume_input

See CreateVolumeInput

delete_backup_input

See DeleteBackupInput

delete_file_system_input

See DeleteFileSystemInput

delete_storage_virtual_machine_input

See DeleteStorageVirtualMachineInput

delete_volume_input

See DeleteVolumeInput

describe_backups_input

See DescribeBackupsInput

describe_data_repository_tasks_input

See DescribeDataRepositoryTasksInput

describe_file_system_aliases_input

See DescribeFileSystemAliasesInput

describe_file_systems_input

See DescribeFileSystemsInput

describe_storage_virtual_machines_input

See DescribeStorageVirtualMachinesInput

describe_volumes_input

See DescribeVolumesInput

disassociate_file_system_aliases_input

See DisassociateFileSystemAliasesInput

list_tags_for_resource_input

See ListTagsForResourceInput

tag_resource_input

See TagResourceInput

untag_resource_input

See UntagResourceInput

update_file_system_input

See UpdateFileSystemInput

update_storage_virtual_machine_input

See UpdateStorageVirtualMachineInput

update_volume_input

See UpdateVolumeInput

Structs

AssociateFileSystemAliasesInput

The request object specifying one or more DNS alias names to associate with an Amazon FSx for Windows File Server file system.

CancelDataRepositoryTaskInput

Cancels a data repository task.

CopyBackupInput

CreateBackupInput

The request object for the CreateBackup operation.

CreateDataRepositoryTaskInput

CreateFileSystemFromBackupInput

The request object for the CreateFileSystemFromBackup

CreateFileSystemInput

The request object used to create a new Amazon FSx file system.

CreateStorageVirtualMachineInput

CreateVolumeFromBackupInput

CreateVolumeInput

DeleteBackupInput

The request object for DeleteBackup operation.

DeleteFileSystemInput

The request object for DeleteFileSystem operation.

DeleteStorageVirtualMachineInput

DeleteVolumeInput

DescribeBackupsInput

The request object for DescribeBackups operation.

DescribeDataRepositoryTasksInput

DescribeFileSystemAliasesInput

The request object for DescribeFileSystemAliases operation.

DescribeFileSystemsInput

The request object for DescribeFileSystems operation.

DescribeStorageVirtualMachinesInput

DescribeVolumesInput

DisassociateFileSystemAliasesInput

The request object of DNS aliases to disassociate from an Amazon FSx for Windows File Server file system.

ListTagsForResourceInput

The request object for ListTagsForResource operation.

TagResourceInput

The request object for the TagResource operation.

UntagResourceInput

The request object for UntagResource action.

UpdateFileSystemInput

The request object for the UpdateFileSystem operation.

UpdateStorageVirtualMachineInput

UpdateVolumeInput