Module aws_sdk_fsx::model

Modules

active_directory_backup_attributes

See ActiveDirectoryBackupAttributes

administrative_action

See AdministrativeAction

administrative_action_failure_details

See AdministrativeActionFailureDetails

alias

See Alias

backup

See Backup

backup_failure_details

See BackupFailureDetails

completion_report

See CompletionReport

create_file_system_lustre_configuration

See CreateFileSystemLustreConfiguration

create_file_system_ontap_configuration

See CreateFileSystemOntapConfiguration

create_file_system_windows_configuration

See CreateFileSystemWindowsConfiguration

create_ontap_volume_configuration

See CreateOntapVolumeConfiguration

create_svm_active_directory_configuration

See CreateSvmActiveDirectoryConfiguration

data_repository_configuration

See DataRepositoryConfiguration

data_repository_failure_details

See DataRepositoryFailureDetails

data_repository_task

See DataRepositoryTask

data_repository_task_failure_details

See DataRepositoryTaskFailureDetails

data_repository_task_filter

See DataRepositoryTaskFilter

data_repository_task_status

See DataRepositoryTaskStatus

delete_file_system_lustre_configuration

See DeleteFileSystemLustreConfiguration

delete_file_system_lustre_response

See DeleteFileSystemLustreResponse

delete_file_system_windows_configuration

See DeleteFileSystemWindowsConfiguration

delete_file_system_windows_response

See DeleteFileSystemWindowsResponse

delete_volume_ontap_configuration

See DeleteVolumeOntapConfiguration

delete_volume_ontap_response

See DeleteVolumeOntapResponse

disk_iops_configuration

See DiskIopsConfiguration

file_system

See FileSystem

file_system_endpoint

See FileSystemEndpoint

file_system_endpoints

See FileSystemEndpoints

file_system_failure_details

See FileSystemFailureDetails

filter

See Filter

lifecycle_transition_reason

See LifecycleTransitionReason

lustre_file_system_configuration

See LustreFileSystemConfiguration

ontap_file_system_configuration

See OntapFileSystemConfiguration

ontap_volume_configuration

See OntapVolumeConfiguration

self_managed_active_directory_attributes

See SelfManagedActiveDirectoryAttributes

self_managed_active_directory_configuration

See SelfManagedActiveDirectoryConfiguration

self_managed_active_directory_configuration_updates

See SelfManagedActiveDirectoryConfigurationUpdates

storage_virtual_machine

See StorageVirtualMachine

storage_virtual_machine_filter

See StorageVirtualMachineFilter

svm_active_directory_configuration

See SvmActiveDirectoryConfiguration

svm_endpoint

See SvmEndpoint

svm_endpoints

See SvmEndpoints

tag

See Tag

tiering_policy

See TieringPolicy

update_file_system_lustre_configuration

See UpdateFileSystemLustreConfiguration

update_file_system_ontap_configuration

See UpdateFileSystemOntapConfiguration

update_file_system_windows_configuration

See UpdateFileSystemWindowsConfiguration

update_ontap_volume_configuration

See UpdateOntapVolumeConfiguration

update_svm_active_directory_configuration

See UpdateSvmActiveDirectoryConfiguration

volume

See Volume

volume_filter

See VolumeFilter

windows_audit_log_configuration

See WindowsAuditLogConfiguration

windows_audit_log_create_configuration

See WindowsAuditLogCreateConfiguration

windows_file_system_configuration

See WindowsFileSystemConfiguration

Structs

ActiveDirectoryBackupAttributes

The Microsoft AD attributes of the Amazon FSx for Windows File Server file system.

AdministrativeAction

Describes a specific Amazon FSx administrative action for the current Windows or

AdministrativeActionFailureDetails

Provides information about a failed administrative action.

Alias

A DNS alias that is associated with the file system. You can use a DNS alias to access a file system using

Backup

A backup of an Amazon FSx for Windows File Server or Amazon FSx for Lustre file system,

BackupFailureDetails

If backup creation fails, this structure contains the details of that

CompletionReport

Provides a report detailing the data repository task results of the files processed that match the criteria specified in the report Scope parameter.

CreateFileSystemLustreConfiguration

The Lustre configuration for the file system being created.

CreateFileSystemOntapConfiguration

The ONTAP configuration properties of the FSx for NetApp ONTAP file system that you are creating.

CreateFileSystemWindowsConfiguration

The configuration object for the Microsoft Windows file system used in

CreateOntapVolumeConfiguration

Specifies the configuration of the ONTAP volume that you are creating.

CreateSvmActiveDirectoryConfiguration

The configuration that Amazon FSx uses to join the ONTAP storage virtual machine

DataRepositoryConfiguration

The data repository configuration object for Lustre file systems returned in the response of

DataRepositoryFailureDetails

Provides detailed information about the data respository if its Lifecycle is

DataRepositoryTask

A description of the data repository task. You use data repository tasks

DataRepositoryTaskFailureDetails

Provides information about why a data repository task failed. Only populated when the task Lifecycle is set to FAILED.

DataRepositoryTaskFilter

(Optional) An array of filter objects you can use to filter the response of data repository tasks you will see in the the response.

DataRepositoryTaskStatus

Provides the task status showing a running total of the total number of files to be processed,

DeleteFileSystemLustreConfiguration

The configuration object for the Amazon FSx for Lustre file system being deleted in the

DeleteFileSystemLustreResponse

The response object for the Amazon FSx for Lustre file system being deleted in the

DeleteFileSystemWindowsConfiguration

The configuration object for the Microsoft Windows file system used in the

DeleteFileSystemWindowsResponse

The response object for the Microsoft Windows file system used in the

DeleteVolumeOntapConfiguration

Use to specify skipping a final backup, or to add tags to a final backup.

DeleteVolumeOntapResponse

The response object for the Amazon FSx for NetApp ONTAP volume being deleted

DiskIopsConfiguration

The SSD IOPS (input/output operations per second) configuration

FileSystem

A description of a specific Amazon FSx file system.

FileSystemEndpoint

An Amazon FSx for NetApp ONTAP file system has two endpoints

FileSystemEndpoints

An Amazon FSx for NetApp ONTAP file system has the following endpoints

FileSystemFailureDetails

A structure providing details of any failures that occur when creating the file system

Filter

A filter used to restrict the results of describe calls. You can use multiple

LifecycleTransitionReason

Describes why a resource lifecycle state changed.

LustreFileSystemConfiguration

The configuration for the Amazon FSx for Lustre file system.

OntapFileSystemConfiguration

Configuration for the FSx for NetApp ONTAP file system.

OntapVolumeConfiguration

The configuration of an Amazon FSx for NetApp ONTAP volume

SelfManagedActiveDirectoryAttributes

The configuration of the self-managed Microsoft Active Directory (AD) directory to

SelfManagedActiveDirectoryConfiguration

The configuration that Amazon FSx uses to join a Amazon FSx for Windows File Server file system or an ONTAP storage virtual machine (SVM) to

SelfManagedActiveDirectoryConfigurationUpdates

The configuration that Amazon FSx uses to join the Windows File Server instance to a

StorageVirtualMachine

Describes the Amazon FSx for NetApp ONTAP storage virtual machine (SVM) configuraton.

StorageVirtualMachineFilter

A filter used to restrict the results of describe calls for

SvmActiveDirectoryConfiguration

Describes the configuration of the Microsoft Active Directory (AD)

SvmEndpoint

An Amazon FSx for NetApp ONTAP storage virtual machine (SVM) has

SvmEndpoints

An Amazon FSx for NetApp ONTAP storage virtual machine (SVM) has

Tag

Specifies a key-value pair for a resource tag.

TieringPolicy

Describes the data tiering policy for an ONTAP volume. When enabled, Amazon FSx for ONTAP's intelligent

UpdateFileSystemLustreConfiguration

The configuration object for Amazon FSx for Lustre file systems used in the

UpdateFileSystemOntapConfiguration

The configuration updates for an Amazon FSx for NetApp ONTAP file system.

UpdateFileSystemWindowsConfiguration

Updates the configuration for an existing Amazon FSx for Windows

UpdateOntapVolumeConfiguration

Used to specify changes to the ONTAP configuration for the volume you are updating.

UpdateSvmActiveDirectoryConfiguration

Updates the Microsoft Active Directory (AD) configuration of an SVM joined to an AD.

Volume

Describes an Amazon FSx for NetApp ONTAP volume.

VolumeFilter

A filter used to restrict the results of describe calls for

WindowsAuditLogConfiguration

The configuration that Amazon FSx for Windows File Server uses to audit and log

WindowsAuditLogCreateConfiguration

The Windows file access auditing configuration used when creating

WindowsFileSystemConfiguration

The configuration for this Microsoft Windows file system.

Enums

ActiveDirectoryErrorType

The type of error relating to Microsoft Active Directory. NOT_FOUND means that no

AdministrativeActionType

Describes the type of administrative action, as follows:

AliasLifecycle

AutoImportPolicyType

BackupLifecycle

The lifecycle status of the backup.

BackupType

The type of the backup.

DataCompressionType

DataRepositoryLifecycle

DataRepositoryTaskFilterName

DataRepositoryTaskLifecycle

DataRepositoryTaskType

DiskIopsConfigurationMode

DriveCacheType

FileSystemLifecycle

The lifecycle status of the file system.

FileSystemMaintenanceOperation

An enumeration specifying the currently ongoing maintenance operation.

FileSystemType

The type of file system.

FilterName

The name for a filter.

FlexCacheEndpointType

LustreDeploymentType

OntapDeploymentType

OntapVolumeType

ReportFormat

ReportScope

ResourceType

SecurityStyle

ServiceLimit

The types of limits on your service utilization. Limits include file system count,

Status

StorageType

The storage type for your Amazon FSx file system.

StorageVirtualMachineFilterName

StorageVirtualMachineLifecycle

StorageVirtualMachineRootVolumeSecurityStyle

StorageVirtualMachineSubtype

TieringPolicyName

VolumeFilterName

VolumeLifecycle

VolumeType

WindowsAccessAuditLogLevel

WindowsDeploymentType