Struct aws_sdk_fsx::model::FileSystem

#[non_exhaustive]
pub struct FileSystem {

    pub owner_id: Option<String>,
    pub creation_time: Option<Instant>,
    pub file_system_id: Option<String>,
    pub file_system_type: Option<FileSystemType>,
    pub lifecycle: Option<FileSystemLifecycle>,
    pub failure_details: Option<FileSystemFailureDetails>,
    pub storage_capacity: Option<i32>,
    pub storage_type: Option<StorageType>,
    pub vpc_id: Option<String>,
    pub subnet_ids: Option<Vec<String>>,
    pub network_interface_ids: Option<Vec<String>>,
    pub dns_name: Option<String>,
    pub kms_key_id: Option<String>,
    pub resource_arn: Option<String>,
    pub tags: Option<Vec<Tag>>,
    pub windows_configuration: Option<WindowsFileSystemConfiguration>,
    pub lustre_configuration: Option<LustreFileSystemConfiguration>,
    pub administrative_actions: Option<Vec<AdministrativeAction>>,
    pub ontap_configuration: Option<OntapFileSystemConfiguration>,

}

A description of a specific Amazon FSx file system.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

owner_id: Option<String>

The Amazon Web Services account that created the file system. If the file system was created by an Identity and Access Management (IAM) user, the Amazon Web Services account to which the IAM user belongs is the owner.

creation_time: Option<Instant>

The time that the file system was created, in seconds (since 1970-01-01T00:00:00Z), also known as Unix time.

file_system_id: Option<String>

The system-generated, unique 17-digit ID of the file system.

file_system_type: Option<FileSystemType>

The type of Amazon FSx file system, which can be LUSTRE, WINDOWS, or ONTAP.

lifecycle: Option<FileSystemLifecycle>

The lifecycle status of the file system, following are the possible values and what they mean:

AVAILABLE - The file system is in a healthy state, and is reachable and available for use.

CREATING - Amazon FSx is creating the new file system.

DELETING - Amazon FSx is deleting an existing file system.

FAILED - An existing file system has experienced an unrecoverable failure. When creating a new file system, Amazon FSx was unable to create the file system.

MISCONFIGURED indicates that the file system is in a failed but recoverable state.

UPDATING indicates that the file system is undergoing a customer initiated update.

failure_details: Option<FileSystemFailureDetails>

A structure providing details of any failures that occur when creating the file system has failed.

storage_capacity: Option<i32>

The storage capacity of the file system in gibibytes (GiB).

storage_type: Option<StorageType>

The storage type of the file system. Valid values are SSD and HDD. If set to SSD, the file system uses solid state drive storage. If set to HDD, the file system uses hard disk drive storage.

vpc_id: Option<String>

The ID of the primary VPC for the file system.

subnet_ids: Option<Vec<String>>

Specifies the IDs of the subnets that the file system is accessible from. For Windows and ONTAP MULTI_AZ_1 file system deployment type, there are two subnet IDs, one for the preferred file server and one for the standby file server. The preferred file server subnet identified in the PreferredSubnetID property. All other file systems have only one subnet ID.

For Lustre file systems, and Single-AZ Windows file systems, this is the ID of the subnet that contains the endpoint for the file system. For MULTI_AZ_1 Windows and ONTAP file systems, the endpoint for the file system is available in the PreferredSubnetID.

network_interface_ids: Option<Vec<String>>

The IDs of the elastic network interface from which a specific file system is accessible. The elastic network interface is automatically created in the same VPC that the Amazon FSx file system was created in. For more information, see Elastic Network Interfaces in the Amazon EC2 User Guide.

For an Amazon FSx for Windows File Server file system, you can have one network interface ID. For an Amazon FSx for Lustre file system, you can have more than one.

dns_name: Option<String>

The DNS name for the file system.

kms_key_id: Option<String>

The ID of the Key Management Service (KMS) key used to encrypt the file system's data for Amazon FSx for Windows File Server file systems, Amazon FSx for NetApp ONTAP file systems, and persistent Amazon FSx for Lustre file systems at rest. If not specified, the Amazon FSx managed key is used. The scratch Amazon FSx for Lustre file systems are always encrypted at rest using Amazon FSx managed keys. For more information, see Encrypt in the Key Management Service API Reference.

resource_arn: Option<String>

The Amazon Resource Name (ARN) for the file system resource.

tags: Option<Vec<Tag>>

The tags to associate with the file system. For more information, see Tagging Your Amazon EC2 Resources in the Amazon EC2 User Guide.

windows_configuration: Option<WindowsFileSystemConfiguration>

The configuration for this Microsoft Windows file system.

lustre_configuration: Option<LustreFileSystemConfiguration>

The configuration for the Amazon FSx for Lustre file system.

administrative_actions: Option<Vec<AdministrativeAction>>

A list of administrative actions for the file system that are in process or waiting to be processed. Administrative actions describe changes to the Amazon FSx file system that you have initiated using the UpdateFileSystem action.

ontap_configuration: Option<OntapFileSystemConfiguration>

The configuration for this FSx for NetApp ONTAP file system.

Implementations

impl FileSystem

pub fn builder() -> Builder

Creates a new builder-style object to manufacture FileSystem