Module aws_sdk_fsx::operation

Structs

AssociateFileSystemAliases

Use this action to associate one or more Domain Name Server (DNS) aliases with an existing Amazon FSx for Windows File Server file system.

CancelDataRepositoryTask

Cancels an existing Amazon FSx for Lustre data repository task if that task is in either the

CopyBackup

Copies an existing backup within the same Amazon Web Services account to another Amazon Web Services Region

CreateBackup

Creates a backup of an existing Amazon FSx for Windows File Server

CreateDataRepositoryTask

Creates an Amazon FSx for Lustre data repository task. You use data repository tasks

CreateFileSystem

Creates a new, empty Amazon FSx file system.

CreateFileSystemFromBackup

Creates a new Amazon FSx for Lustre or Amazon FSx for Windows File Server file system

CreateStorageVirtualMachine

Creates a storage virtual machine (SVM) for an Amazon FSx for ONTAP file system.

CreateVolume

Creates an Amazon FSx for NetApp ONTAP storage volume.

CreateVolumeFromBackup

Creates a new Amazon FSx for NetApp ONTAP volume from an

DeleteBackup

Deletes an Amazon FSx backup, deleting its contents. After deletion, the backup no longer exists, and its data is gone.

DeleteFileSystem

Deletes a file system, deleting its contents. After deletion, the file system no

DeleteStorageVirtualMachine

Deletes an existing Amazon FSx for ONTAP storage virtual machine (SVM). Prior

DeleteVolume

Deletes an Amazon FSx for NetApp ONTAP volume. When deleting a volume,

DescribeBackups

Returns the description of specific Amazon FSx backups, if

DescribeDataRepositoryTasks

Returns the description of specific Amazon FSx for Lustre data repository tasks, if

DescribeFileSystemAliases

Returns the DNS aliases that are associated with the specified Amazon FSx for Windows File Server file system. A history of

DescribeFileSystems

Returns the description of specific Amazon FSx file systems, if a

DescribeStorageVirtualMachines

Describes one or more Amazon FSx for NetApp ONTAP storage virtual machines (SVMs).

DescribeVolumes

Describes one or more Amazon FSx for NetApp ONTAP volumes.

DisassociateFileSystemAliases

Use this action to disassociate, or remove, one or more Domain Name Service (DNS) aliases

ListTagsForResource

Lists tags for an Amazon FSx file systems and backups in the case of Amazon FSx for

TagResource

Tags an Amazon FSx resource.

UntagResource

This action removes a tag from an Amazon FSx resource.

UpdateFileSystem

Use this operation to update the configuration of an existing Amazon FSx file system.

UpdateStorageVirtualMachine

Updates an Amazon FSx for ONTAP storage virtual machine (SVM).

UpdateVolume

Updates an Amazon FSx for NetApp ONTAP volume's configuration.