Module aws_sdk_fsx::output

Modules

associate_file_system_aliases_output

See AssociateFileSystemAliasesOutput

cancel_data_repository_task_output

See CancelDataRepositoryTaskOutput

copy_backup_output

See CopyBackupOutput

create_backup_output

See CreateBackupOutput

create_data_repository_task_output

See CreateDataRepositoryTaskOutput

create_file_system_from_backup_output

See CreateFileSystemFromBackupOutput

create_file_system_output

See CreateFileSystemOutput

create_storage_virtual_machine_output

See CreateStorageVirtualMachineOutput

create_volume_from_backup_output

See CreateVolumeFromBackupOutput

create_volume_output

See CreateVolumeOutput

delete_backup_output

See DeleteBackupOutput

delete_file_system_output

See DeleteFileSystemOutput

delete_storage_virtual_machine_output

See DeleteStorageVirtualMachineOutput

delete_volume_output

See DeleteVolumeOutput

describe_backups_output

See DescribeBackupsOutput

describe_data_repository_tasks_output

See DescribeDataRepositoryTasksOutput

describe_file_system_aliases_output

See DescribeFileSystemAliasesOutput

describe_file_systems_output

See DescribeFileSystemsOutput

describe_storage_virtual_machines_output

See DescribeStorageVirtualMachinesOutput

describe_volumes_output

See DescribeVolumesOutput

disassociate_file_system_aliases_output

See DisassociateFileSystemAliasesOutput

list_tags_for_resource_output

See ListTagsForResourceOutput

tag_resource_output

See TagResourceOutput

untag_resource_output

See UntagResourceOutput

update_file_system_output

See UpdateFileSystemOutput

update_storage_virtual_machine_output

See UpdateStorageVirtualMachineOutput

update_volume_output

See UpdateVolumeOutput

Structs

AssociateFileSystemAliasesOutput

The system generated response showing the DNS aliases that

CancelDataRepositoryTaskOutput

CopyBackupOutput

CreateBackupOutput

The response object for the CreateBackup operation.

CreateDataRepositoryTaskOutput

CreateFileSystemFromBackupOutput

The response object for the CreateFileSystemFromBackup

CreateFileSystemOutput

The response object returned after the file system is created.

CreateStorageVirtualMachineOutput

CreateVolumeFromBackupOutput

CreateVolumeOutput

DeleteBackupOutput

The response object for DeleteBackup operation.

DeleteFileSystemOutput

The response object for the DeleteFileSystem operation.

DeleteStorageVirtualMachineOutput

DeleteVolumeOutput

DescribeBackupsOutput

Response object for DescribeBackups operation.

DescribeDataRepositoryTasksOutput

DescribeFileSystemAliasesOutput

The response object for DescribeFileSystemAliases operation.

DescribeFileSystemsOutput

The response object for DescribeFileSystems operation.

DescribeStorageVirtualMachinesOutput

DescribeVolumesOutput

DisassociateFileSystemAliasesOutput

The system generated response showing the DNS aliases that Amazon FSx is attempting to disassociate from the file system.

ListTagsForResourceOutput

The response object for ListTagsForResource operation.

TagResourceOutput

The response object for the TagResource operation.

UntagResourceOutput

The response object for UntagResource action.

UpdateFileSystemOutput

The response object for the UpdateFileSystem operation.

UpdateStorageVirtualMachineOutput

UpdateVolumeOutput