Module aws_sdk_fsx::model

Modules

active_directory_backup_attributes

See ActiveDirectoryBackupAttributes

administrative_action

See AdministrativeAction

administrative_action_failure_details

See AdministrativeActionFailureDetails

alias

See Alias

backup

See Backup

backup_failure_details

See BackupFailureDetails

completion_report

See CompletionReport

create_file_system_lustre_configuration

See CreateFileSystemLustreConfiguration

create_file_system_windows_configuration

See CreateFileSystemWindowsConfiguration

data_repository_configuration

See DataRepositoryConfiguration

data_repository_failure_details

See DataRepositoryFailureDetails

data_repository_task

See DataRepositoryTask

data_repository_task_failure_details

See DataRepositoryTaskFailureDetails

data_repository_task_filter

See DataRepositoryTaskFilter

data_repository_task_status

See DataRepositoryTaskStatus

delete_file_system_lustre_configuration

See DeleteFileSystemLustreConfiguration

delete_file_system_lustre_response

See DeleteFileSystemLustreResponse

delete_file_system_windows_configuration

See DeleteFileSystemWindowsConfiguration

delete_file_system_windows_response

See DeleteFileSystemWindowsResponse

file_system

See FileSystem

file_system_failure_details

See FileSystemFailureDetails

filter

See Filter

lustre_file_system_configuration

See LustreFileSystemConfiguration

self_managed_active_directory_attributes

See SelfManagedActiveDirectoryAttributes

self_managed_active_directory_configuration

See SelfManagedActiveDirectoryConfiguration

self_managed_active_directory_configuration_updates

See SelfManagedActiveDirectoryConfigurationUpdates

tag

See Tag

update_file_system_lustre_configuration

See UpdateFileSystemLustreConfiguration

update_file_system_windows_configuration

See UpdateFileSystemWindowsConfiguration

windows_audit_log_configuration

See WindowsAuditLogConfiguration

windows_audit_log_create_configuration

See WindowsAuditLogCreateConfiguration

windows_file_system_configuration

See WindowsFileSystemConfiguration

Structs

ActiveDirectoryBackupAttributes

The Microsoft AD attributes of the Amazon FSx for Windows File Server file system.

AdministrativeAction

Describes a specific Amazon FSx administrative action for the current Windows or

AdministrativeActionFailureDetails

Provides information about a failed administrative action.

Alias

A DNS alias that is associated with the file system. You can use a DNS alias to access a file system using

Backup

A backup of an Amazon FSx file system.

BackupFailureDetails

If backup creation fails, this structure contains the details of that

CompletionReport

Provides a report detailing the data repository task results of the files processed that match the criteria specified in the report Scope parameter.

CreateFileSystemLustreConfiguration

The Lustre configuration for the file system being created.

CreateFileSystemWindowsConfiguration

The configuration object for the Microsoft Windows file system used in

DataRepositoryConfiguration

The data repository configuration object for Lustre file systems returned in the response of

DataRepositoryFailureDetails

Provides detailed information about the data respository if its Lifecycle is

DataRepositoryTask

A description of the data repository task. You use data repository tasks

DataRepositoryTaskFailureDetails

Provides information about why a data repository task failed. Only populated when the task Lifecycle is set to FAILED.

DataRepositoryTaskFilter

(Optional) An array of filter objects you can use to filter the response of data repository tasks you will see in the the response.

DataRepositoryTaskStatus

Provides the task status showing a running total of the total number of files to be processed,

DeleteFileSystemLustreConfiguration

The configuration object for the Amazon FSx for Lustre file system being deleted in the

DeleteFileSystemLustreResponse

The response object for the Amazon FSx for Lustre file system being deleted in the

DeleteFileSystemWindowsConfiguration

The configuration object for the Microsoft Windows file system used in the

DeleteFileSystemWindowsResponse

The response object for the Microsoft Windows file system used in the

FileSystem

A description of a specific Amazon FSx file system.

FileSystemFailureDetails

A structure providing details of any failures that occur when creating the file system

Filter

A filter used to restrict the results of describe calls. You can use multiple

LustreFileSystemConfiguration

The configuration for the Amazon FSx for Lustre file system.

SelfManagedActiveDirectoryAttributes

The configuration of the self-managed Microsoft Active Directory (AD) directory to

SelfManagedActiveDirectoryConfiguration

The configuration that Amazon FSx uses to join the Windows File Server instance to

SelfManagedActiveDirectoryConfigurationUpdates

The configuration that Amazon FSx uses to join the Windows File Server instance to a

Tag

Specifies a key-value pair for a resource tag.

UpdateFileSystemLustreConfiguration

The configuration object for Amazon FSx for Lustre file systems used in the

UpdateFileSystemWindowsConfiguration

Updates the configuration for an existing Amazon FSx for Windows

WindowsAuditLogConfiguration

The configuration that Amazon FSx for Windows File Server uses to audit and log

WindowsAuditLogCreateConfiguration

The Windows file access auditing configuration used when creating

WindowsFileSystemConfiguration

The configuration for this Microsoft Windows file system.

Enums

ActiveDirectoryErrorType

The type of error relating to Microsoft Active Directory. NOT_FOUND means that no

AdministrativeActionType

Describes the type of administrative action, as follows:

AliasLifecycle

AutoImportPolicyType

BackupLifecycle

The lifecycle status of the backup.

BackupType

The type of the backup.

DataCompressionType

DataRepositoryLifecycle

DataRepositoryTaskFilterName

DataRepositoryTaskLifecycle

DataRepositoryTaskType

DriveCacheType

FileSystemLifecycle

The lifecycle status of the file system.

FileSystemMaintenanceOperation

An enumeration specifying the currently ongoing maintenance operation.

FileSystemType

The type of file system.

FilterName

The name for a filter.

LustreDeploymentType

ReportFormat

ReportScope

ServiceLimit

The types of limits on your service utilization. Limits include file system count,

Status

StorageType

The storage type for your Amazon FSx file system.

WindowsAccessAuditLogLevel

WindowsDeploymentType