Module aws_sdk_fsx::operation

Structs

AssociateFileSystemAliases

Use this action to associate one or more Domain Name Server (DNS) aliases with an existing Amazon FSx for Windows File Server file system.

CancelDataRepositoryTask

Cancels an existing Amazon FSx for Lustre data repository task if that task is in either the

CopyBackup

Copies an existing backup within the same AWS account to another Region

CreateBackup

Creates a backup of an existing Amazon FSx file system.

CreateDataRepositoryTask

Creates an Amazon FSx for Lustre data repository task. You use data repository tasks

CreateFileSystem

Creates a new, empty Amazon FSx file system.

CreateFileSystemFromBackup

Creates a new Amazon FSx file system from an existing Amazon FSx backup.

DeleteBackup

Deletes an Amazon FSx backup, deleting its contents. After deletion, the backup no longer exists, and its data is gone.

DeleteFileSystem

Deletes a file system, deleting its contents. After deletion, the file system no

DescribeBackups

Returns the description of specific Amazon FSx backups, if

DescribeDataRepositoryTasks

Returns the description of specific Amazon FSx for Lustre data repository tasks, if

DescribeFileSystemAliases

Returns the DNS aliases that are associated with the specified Amazon FSx for Windows File Server file system. A history of

DescribeFileSystems

Returns the description of specific Amazon FSx file systems, if a

DisassociateFileSystemAliases

Use this action to disassociate, or remove, one or more Domain Name Service (DNS) aliases

ListTagsForResource

Lists tags for an Amazon FSx file systems and backups in the case of Amazon FSx for

TagResource

Tags an Amazon FSx resource.

UntagResource

This action removes a tag from an Amazon FSx resource.

UpdateFileSystem

Use this operation to update the configuration of an existing Amazon FSx file system.